# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WK API",
    "description": "We Kare Api 1.0 ",
    "termsOfService": "http://www.wkya.rw",
    "contact": {
      "name": "wecare",
      "url": "http://www.wkya.rw",
      "email": "rwahamanick@gmail.com"
    },
    "version": "v1",
    "x-logo": {
      "url": "/img/tiger-soft-logo.png",
      "backgroundColor": "#FFFFFF",
      "altText": "PetStore Logo"
    }
  },
  "paths": {
    "/barbers": {
      "get": {
        "tags": [
          "Barber"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Barber"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/barbers/partner/{partnerId}": {
      "get": {
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/barbers/{id}": {
      "get": {
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Barber"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/businessClients": {
      "get": {
        "tags": [
          "BusinessClient"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "BusinessClient"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusinessClientInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/businessClients/{id}": {
      "get": {
        "tags": [
          "BusinessClient"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "BusinessClient"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusinessClientInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BusinessClient"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/businessClients/serialNumber/{serialNumber}": {
      "get": {
        "tags": [
          "BusinessClient"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serialNumber",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservations/client/{clientId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BarberInputDto": {
      "required": [
        "names",
        "phone"
      ],
      "type": "object",
      "properties": {
        "names": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "partnerId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "BusinessClientInputDto": {
      "required": [
        "name",
        "phone"
      ],
      "type": "object",
      "properties": {
        "phone": {
          "maxLength": 10,
          "minLength": 10,
          "type": "string"
        },
        "name": {
          "maxLength": 125,
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "serialNumber": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Authorization"
    }
  },
  "security": [
    { }
  ]
}